Renaming ResponseExample back to Example.
