+++ packages/api-explorer/__tests__/Example.test.jsx
@@ -5,7 +5,7 @@ const string = require('./fixtures/string/oas.json');
 const exampleResults = require('./fixtures/example-results/oas');
 const extensions = require('@readme/oas-extensions');
 
-const ResponseExample = require('../src/ResponseExample');
+const Example = require('../src/Example');
 const Oas = require('../src/lib/Oas');
 
 const oas = new Oas(petstore);
@@ -19,14 +19,14 @@ const props = {
 };
 
 test('should show no examples if endpoint does not any', () => {
-  const example = shallow(<ResponseExample {...props} />);
+  const example = shallow(<Example {...props} />);
 
   expect(example.containsMatchingElement(<div>Try the API to see Results</div>)).toEqual(true);
 });
 
 test('should notify about no examples being available if explorer disabled', () => {
   const example = shallow(
-    <ResponseExample
+    <Example
       {...props}
       oas={new Oas(Object.assign({}, petstore, { [extensions.EXPLORER_ENABLED]: false }))}
     />,
@@ -38,11 +38,7 @@ test('should notify about no examples being available if explorer disabled', () 
 test('should show each example', () => {
   const exampleOas = new Oas(exampleResults);
   const example = shallow(
-    <ResponseExample
-      {...props}
-      oas={exampleOas}
-      operation={exampleOas.operation('/results', 'get')}
-    />,
+    <Example {...props} oas={exampleOas} operation={exampleOas.operation('/results', 'get')} />,
   );
 
   expect(example.find('pre').length).toEqual(2);
@@ -51,11 +47,7 @@ test('should show each example', () => {
 test('should display json viewer', () => {
   const exampleOas = new Oas(exampleResults);
   const example = shallow(
-    <ResponseExample
-      {...props}
-      oas={exampleOas}
-      operation={exampleOas.operation('/results', 'get')}
-    />,
+    <Example {...props} oas={exampleOas} operation={exampleOas.operation('/results', 'get')} />,
   );
 
   // Asserting all JSON examples are displayed with JSON viewer from the example oas.json
@@ -71,11 +63,7 @@ test('should display json viewer', () => {
 test('should not fail to parse invalid json and instead show the standard syntax highlighter', () => {
   const exampleOas = new Oas(string);
   const example = shallow(
-    <ResponseExample
-      {...props}
-      oas={exampleOas}
-      operation={exampleOas.operation('/format-uuid', 'get')}
-    />,
+    <Example {...props} oas={exampleOas} operation={exampleOas.operation('/format-uuid', 'get')} />,
   );
 
   // Asserting that instead of failing with the invalid JSON we attempted to render, we fallback
@@ -92,11 +80,7 @@ test('should not fail to parse invalid json and instead show the standard syntax
 test('should correctly highlight XML syntax', () => {
   const exampleOas = new Oas(exampleResults);
   const example = shallow(
-    <ResponseExample
-      {...props}
-      oas={exampleOas}
-      operation={exampleOas.operation('/results', 'get')}
-    />,
+    <Example {...props} oas={exampleOas} operation={exampleOas.operation('/results', 'get')} />,
   );
 
   // Asserting that there are XML tags
@@ -112,7 +96,7 @@ test('should correctly highlight XML syntax', () => {
 test('should show select for multiple response types', () => {
   const exampleOas = new Oas(exampleResults);
   const example = shallow(
-    <ResponseExample
+    <Example
       {...props}
       oas={exampleOas}
       operation={exampleOas.operation('/multi-media-types', 'get')}
@@ -129,7 +113,7 @@ test('should show select for multiple response types', () => {
 test('should show select for multiple response type examples', () => {
   const exampleOas = new Oas(exampleResults);
   const example = shallow(
-    <ResponseExample
+    <Example
       {...props}
       oas={exampleOas}
       operation={exampleOas.operation('/multi-results', 'get')}
